requirements.txt
Django==1.7
Markdown==2.4
Pygments==1.6

coverage==3.7.1
django-appconf==0.6
django-compressor==1.4
django-tinymce==1.5.2
hamlpy==0.82.2
six==1.6.1
wsgiref==0.1.2
django-sekizai
djangorestframework
redis
